19111_CustomFieldResolver_Annotations_and_Interface
Move Custom field handling to a seperate class
